confirmation dialog

+++ src/app/components/settings/settings.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { TodoService } from 'src/app/services/todo.service';
+import { ConfirmationDialogComponent } from '../confirmation-dialog/confirmation-dialog.component';
+import { MatDialog } from '@angular/material/dialog';
 
 @Component({
   selector: 'app-settings',
@@ -10,18 +12,25 @@ export class SettingsComponent implements OnInit {
 
   toggleTheme: any;
 
-  constructor(private todoService: TodoService) { }
+  constructor(private todoService: TodoService, private dialog: MatDialog) { }
 
   ngOnInit() {
   }
 
   onChange(event: any) {
-    console.log(event)
+    console.log(event.target.checked)
   }
 
   onDelete(){
-    alert('are you sure you want to delete all tasks? This operation cannot be undone.')
-    this.todoService.deleteAllTasksfromLocalStorage();
+    const confirmDialog = this.dialog.open(ConfirmationDialogComponent,{
+      width: '420px',
+      autoFocus: false
+    })
+    confirmDialog.afterClosed().subscribe(result => {
+      if(result === true) {
+        this.todoService.deleteAllTasksfromLocalStorage();
+      }
+    });
   }
 
 }
